feat(packaging): expose evaluator CLI entry points
Move canonical evaluator commands into atrex_bench.cli and retain compatibility wrappers under scripts/. Publish the three evaluator console commands.

Preserve public package metadata and omit internal publishing infrastructure.
