+++ src/atrex_bench/cli/__init__.py
@@ -1,0 +1,1 @@
+"""Installed command-line entrypoints for Atrex-Bench."""
